# Conflicts:
#   database/models/mp_models/MarketOrder.js
#   src/mpModules/marketSell/index.js
#   src/mpModules/marketSell/marketSellContant.js
#   src/mpModules/marketSell/marketSellController.js
#   src/mpModules/marketSell/marketSellService.js

+++ database/models/mp_models/Transaction.js
@@ -32,7 +32,7 @@ module.exports = (sequelize, DataTypes) => {
                 allowNull: false
             },
             value: {
-                type: DataTypes.INTEGER(11).UNSIGNED,
+                type: DataTypes.BIGINT.UNSIGNED,
                 allowNull: false
             },
             //Tài khoản tạo (tài khoản đăng nhập)
